Gx code cleanup. Licence cleanup. Fixing sub analysis
