+++ cp/Makefile
@@ -76,7 +76,7 @@ SRCS-y += $(SRCDIR)/../test/simu_cp/simu_cp.o
 SRCS-y += $(SRCDIR)/../cp_dp_api/vepc_cp_dp_api.o
 SRCS-y += $(SRCDIR)/../oss_adapter/c3po_oss/gw_adapter.o
 SRCS-y += gx_event_handlers.c 
-SRCS-y += gx_app_interface.c
+SRCS-y += gx_interface.c
 SRCS-y += ipc_api.c
 SRCS-y += state_machine/sm_gx_pcnd.o
 SRCS-y += $(SRCDIR)/gx_app/src/gx_pack.o
