+++ cp/gx_event_handlers.h
@@ -6,7 +6,7 @@
 #define _GX_EVENT_HANDLERS_H
 
 #include "sm_struct.h"
-#include "gx_app_interface.h"
+#include "gx_interface.h"
 
 int process_gx_message(gx_msg *gx_msg, msg_info *msg);
