+++ pfcp_messages/cp/pfcp_cp_session.h
@@ -372,9 +372,6 @@ int
 fill_update_pfcp_info(pfcp_sess_mod_req_t *pfcp_sess_mod_req, dynamic_rule_t *dyn_rule);
 
 int
-gen_reauth_response(ue_context_t *context, uint8_t ebi_index);
-
-int
 fill_remove_pfcp_info(pfcp_sess_mod_req_t *pfcp_sess_mod_req, eps_bearer_t *bearer);
 
 int8_t
